CsipSetCoordinatorService: Fix handling multiple connect/disconnect in SM
This CL make sure that Connect/Disconnect flow will not be disturbed
when those messages are deferred.
In particular, this CL does not allow to put more than single Connect or
Disconnect message on the deferred queue.

This is an improvement for previous CL created based on the unit tests

Bug: 409666634
Bug: 429143534
Test: atest CsipSetCoordinatorStateMachineTest
Flag: com.android.bluetooth.flags.ignore_multiple_connect_request_in_bt_services
Change-Id: Ibbafc8f472c2a705f4122a5c117bce7e7d3e8cd4