packetgen: Fix out of bound access when parsing scalar fields
When parsing scalar fields of non-standard octet size (e.g. 24)
the generator code would access bytes _outside_ the encoded range
in order to apply endianess.

If this field is places at the end of a packet, this is generating
an unchcked out of bound access.

This patch is limited in scope to scalar fields to address the
needs of some HCI packets.

This change cannot be flagged as it modifies a build tool.

Bug: 430842231
Flag: EXEMPT, infeasible
Test: atest --host bluetooth_packet_parser_test
Change-Id: I22d775315269a4460179d0ad439d9e69434e9094