SystemServer On to bleOn via OFF
Doing a ON -> BLE_ON transition prevents a clean reset of the bluetooth
states.

The current workaround is to activate the airplane mode at the same
time, but that not how it should be done.
Some 3p apps even educate their user on this.

With this patch, the ON -> BLE_ON transition will always go to OFF, and
then evaluate if there is a need for BLE_ON. This will start a fresh new
process.

This long awaited patch is now possible thanks to all the previous start
and restart fix that stabilized the bluetooth boot

Bug: 325048421
Bug: 431079894
Test: atest ServiceBluetoothTests
Flag: com.android.bluetooth.flags.on_to_ble_on_via_off
Change-Id: Ib9ca54573365be86167b4217a963e4e7d60ee6ea