AutoOn as member of Bms
In order to prepare for concurrent multi BluetoothManagerService, we
need to encapsulate the dependency as a class.

This CL take care of the AutoOn, without changing the behavior

Also drop the 'feature' out of the name.
Remove 'User' out of method name as is is implied that the AutoOn is a
per user feature entirely

Flag: EXEMPT refactor no-op
Test: atest ServiceBluetoothRoboTests
Bug: 391437857
Change-Id: I9b2d71799c21dde8433a6ab5382036cf79c1c123