Setup Bluetooth component independently of the user
The Bluetooth Component is depending on the device, not on which user is
running.
Current implementation has race condition due to the user being
deletable at any time, and if this happen while we expect to know the
bluetooth name, it can crash.

Instead, we should deduce the name only once at boot, and inject it
where it is needed like when we are dealing with user restriction

Bug: 420992261
Flag: com.android.bluetooth.flags.user_restriction_refactor
Test: atest ServiceBluetoothTests
Change-Id: Ie69e032865e9c76a1cd5375055654beba212b12f