Rename `mStats` to `appScanStats` in ScanClient
The name `mStats` is too ambiguous and simply misleading. This field
does not just store "statistics" of a scan but more so of the
Application that the scan belongs to and all its related scan tracking;
Hence the more accurate name of the field of type: `AppScanStats`

`AppScanStats` is itself also not a 100% accurate name as well but
that's work for another day.

Due to `AppScanStats.java` package-private visibility in Java, the field
has to marked as `internal`. Due to `internal`, when called from Java,
Kotlin adds name mangling resulting in something like
`appScanStats$BluetoothApp`.

The solution, other than exposing `AppScanStats` as public, is to use
`@get:JvmName` and `@set:JvmName`.

For the future, with increased Kotlin usage, this should go down. As
other Kotlin classes can access `internal` fields directly with the
field name as `ScanUtil` is doing in this CL.

The field is used ~80% within Java and ~20% within Kotlin. Due to this,
it is currently left as a JAva optional to provide the safety to Java
but as this ratio grows in Kotlin's favor, converting it to a native
Kotlin optional will make more sense.

Bug: 429793161
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT refactor no-op
Change-Id: Ida9764e1fd5307ed9b82b0373d797be16494966f