Avoid reusing local variables in dumpsys scan duration calculations
This was causing wrong values to be calculated. For example, the
condition to set `suspendDuration = currTime - scan.suspendStartTime`
was guarded by `if (scan.isSuspended) {` so if the current scan was not
suspended, the suspended duration of the previous scan was being added
in `totalSuspendTime += suspendDuration`.

Using `final` in local fields to inline the values and avoid using
new primitives in the for loop.

Also include the number of ongoing scans in output

Bug: 431722194
Bug: 428791432
Test: m BluetoothJavaUnitTests
Flag: EXEMPT dumpsys only
Change-Id: Ie461a114e7668a8937b953ce14304b072d71c1f5