Increase crash recovery timing
In some case, Bluetooth try to recover from a crash too quickly. If that
happen, it server not purpose to spam restart Bluetooth and delaying can
only help.

Instead of a linear increase of time, we are just drastically increasing
the next restart attempt.

Before it was: 400 800 1_200  1_600  2_000
It will now be 400 800 1_200 16_000 20_000

The total time of recovery is increased by 35s, without impacting the
first few attempts in case a fast recover is still possible

(this CL also add a sneaky test for satellite mode, don't mind me ^^)

Bug: 391248793
Test: atest ServiceBluetoothTests
Flag: EXEMPT trivial and 100% coverage tested
Change-Id: Ib9b8e2666d495e6e6de1886e9b3bd3bf36dad0ac