Fix use after free error when calling UnregisterSync
std::promise when passed to the handler (as statically allocated), as soon as it gets finished (on a different thread), the blocking call to future.wait_for() will goes through and UnregisterSync returns. And when the passed closure gets executed, it gets freed as well and while destructing, it crashes.
Converted that to shared_ptr to avoid that.

Bug: 425002597
Bug: 432148110
Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.use_shared_promise_for_le_address_manager
Change-Id: I612deea6692a16cf75607fe572a440e348a9aefa