Improve scan duration tracking in AppScanStats
Previously, scan duration was stored in a `duration` field, which would
lead to inconsistencies, especially with suspended scans. This change
introduces `mEndTimestamp` to `LastScan`, allowing the duration to be
calculated dynamically as `mEndTimestamp - mStartTimestamp`. This
ensures that the less state-full duration calculation always reflects
the actual time between start and stop.

Bug: 431722194
Test: atest BluetoothJavaUnitTests
Test: Manual via dumpsys
Flag: EXEMPT dumpsys only
Change-Id: I4534356263b5ebebc526cd036ff12d3959a7d69f