Only broadcastIntent to concerned user
The Bluetooth app is now started on the correct foreground user. There
is no point in sending intent to all users, as we only want to inform
the current user of the new information.

Note that prior to this change, a fair number of intent were already
sent just for the current user.

Test: m .
Bug: 391437857
Bug: 435244086
Flag: com.android.bluetooth.flags.only_broadcast_to_local_user
Change-Id: I49a33223f50fcf9b98871aa6f199e5176966f3ee