VC native: fix read multiple variable length partitioning
Length in response takes two octets, not one. This patch fixes this,
avoiding a situation that for low MTU values some characteristics
are attempted to be read, but do not fit in the response.

This follows the Table 3.29: Format of the Length Value Tuple in Core
Specification.

Bug: 430458463
Test: atest VolumeControlDeviceTest && atest pts-bot:VCP
Flag: EXEMPT, day-to-day bugfix covered with unit test
Change-Id: Ia4ad8487ae3fcb9fc7297c1e53c6c4bb5cb5eb01