Bass: Simplify the use of SET_BCAST_CODE
The SET_BCAST_CODE message in the state machine is used either to store
provided metadata in mSetBroadcastPINMetadata or to send BROADCAST Code
to a remote device.

It seems that the mechanism used for storing it is redundant and just
complicates the code, and it turns out we can simply remove
mSetBroadcastPINMetadata and mSetBroadcastCodePending.

The API makes sure that the broadcast code is always provided along
with all the metadata, and there is no need to keep it separately.

Bug: 434683497
Bug: 434681485
Flag: com.android.bluetooth.flags.leaudio_broadcast_simplify_set_bcast_code
Test: atest BassClientServiceTest BassClientStateMachineTest
Change-Id: I2a4629d9ca639caf157437011cd5311832e1f246