Ignore scan response when MSFT is enabled
When MSFT is enabled and using a scan filter,
the controller could drop a scan response
if it doesn't match the filter.

To alleviate this, our stack delivers the advertising report
without a scan response, as if it were a passive scan.

There was a bug where we didn't handle this case correctly.
With this change, we monitor if MSFT is enabled and
deliver the advertising report without waiting for a scan response.

Flag: EXEMPT, msft specific change
Bug: 428765273
Test: m -j;
Change-Id: Id899f53eeb20d2567c03fed36daed68f565d2209