gatt: Gatt service should always keep the link
It always makes sense to finish started GATT Discovery session and for
this reason, Gatt Discovery client should not be oportunistic in any
case but instead should keep the link alive.

With this CL we are solving also issue with hanging LE ACL after bonding
of dual mode device which is not on the Allow LeAudio list. By making
Gatt discovery non-oportunistic service, it will trigger inactivity
timer once discovery is completed after bonding which makes sure that ACL is dropped when not needed.
Without this change there were no ACL use which would keep ACL and this is the reason that inactivity timer was never started.

Bug: 383497588
Bug: 436539073
Test: manual testing with Dual mode devices
Test: m packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.gatt_discovery_is_non_opportunistic_client

Change-Id: I56bb0c88adc48ce949800816e9618facb6a6e128