Don't perform CTKD for temporary pairing
BR/EDR to LE CTKD is performed after the link is encrypted. However, the link may get encrypted a bit later after the pairing is already concluded. This leaves a small window for another pairing to be initiated before the prior link is encrypted. As a result, the pairing state holding remote device's security requirements may also change. So CTKD must not be performed based on pairing state, rather on the basis of whether the pairing is persistent or not.

Test: atest BumbleBluetoothTests:android.bluetooth.PairingTest
Flag: com.android.bluetooth.flags.avoid_ctkd_for_temp_pairing
Bug: 423483923
Bug: 436957985
Change-Id: Id0fd18c350b2a589d97b9723ba34178152f463a2