Cherrypick for LE address manager fix.
Bug: 437790650
Fix: 437790650
Flag: EXEMPT, trivial
Test: m Bluetooth

[BLE] BUG_GATT_CLIENT_1

In the loop `for (auto client : registered_clients_)`, the `client` variable was using a copy of the entry, so any changes made to `client` were not reflected.
To directly use the values, the loop was changed to `for (auto& client : registered_clients_)`.

Bug: https://partnerissuetracker.corp.google.com/issues/437790650[BLE] BUG_GATT_CLIENT_1

In the loop `for (auto client : registered_clients_)`, the `client` variable was using a copy of the entry, so any changes made to `client` were not reflected.
To directly use the values, the loop was changed to `for (auto& client : registered_clients_)`.

Bug: https://partnerissuetracker.corp.google.com/issues/437790650
(cherry picked from https://partner-android-review.googlesource.com/q/commit:32cda00f5b596059a3e0685fa747aaf26bb10105)
Merged-In: Iab5287b63b8d268c903c16eef87e0df2469ae809
Change-Id: Iab5287b63b8d268c903c16eef87e0df2469ae809