le_scan: stop all LE scanners if system is suspended
Refactor ScanManager to monitor display changes from AdapterSuspend.

Whenever the screen status changes, the existing logic to handle screen
status is called from the ScanManager. If the system does not go to
sleep after screen is turned off, there is no behavior change.

If the system goes to sleep, AdapterSuspend will call
onSystemSuspendChanged, which stops all regular and batch scans.

When screen is turned on, all scanners are re-enabled. There is no
special handling for system resume, except the systemSuspended status is
updated.

When system is suspended, requests to start scan are queued and put to
the suspended queue.

Finally define a system property to determine if all LE scanners should
be suspended when the system is suspended.

Bug: 438826539
Bug: 438561761
Test: m -j
Test: atest ScanManagerTest
Flag: com.android.bluetooth.flags.stop_le_scan_system_suspend
Change-Id: Ifd5c6c35f77fa79e42e1294226507506fdfdb3b7