le_swoff: Update scheduling priority
Configure the scheduling priority of the
encoder worker thread to SCHED_FIFO with
priority 1.

This thread is timing dependent as a new
encoded frame must be pushed to the
controller every 20ms interval.

Some delay can be absorbed by the
controller buffers but too much jitter
and the controller starts losing packets.

Bug: 438399726
Test: m android.hardware.bluetooth.audio-impl
Test: m libbluetooth_audio_session_aidl
Test: Manual | Update audio HAL and manual streaming test
Flag: EXEMPT, minor bugfix
Change-Id: I0f0ce67d2e9f15eb7e4492c0f1ef88a6eb5f5eec