HIDH: Reset HID CID and state when removing a non-connected device
When setting up HID connection, we already fill in some variables, i.e.
the state and CID of the connection. If at this point we remove the
device, we don't clean those information properly since HID is not
recognized as connected.

Reset the information anyway when a HID device is removed in a
non-connected state.

Bug: 434173919
Bug: 441425948
Test: m com.google.android.bt
Flag: com.android.bluetooth.flags.reset_state_when_removing_non_connected_hid_device
Change-Id: Ifc2d0d22ccaa5daa52f91b059c63a7f3c51a1479