Fork RequiresPermissionDetector from tools/lint_checks/global/checks
This new lint will be very difficult to enable globally as is.
Including the incorrect errors (probably in the few thousands), it
currently yields to ~15k errors.

The current decision is to have the development and rollout first in
the Bluetooth project. Once everything is set, it can either be pushed
back upstream to the global checks or open for enabling to other
individual projects as opt-in.

The only changes from tools/lint_checks/global/checks are:

- Extracting some util constants and helper functions to `Utils.kt`
- package name of main file and test file to:
  - com.google.android.lint.aidl ->
    com.android.bluetooth.lint
- Rename `BINDER_CLASS` -> `CLASS_BINDER` (which is being upstreamed in
ag/35434797)

Everything else remains as is the original files.

Bug: 409768276
Test: atest --host BluetoothLintCheckerTest:com.android.bluetooth.lint.test.RequiresPermissionDetectorTest
Flag: EXEMPT new lint
Change-Id: Iadb9814dd25d53307824dc3bd696345dcff7c6d2