Increase timeouts for GattClientTest#connectAndDisconnectManyClientsWithoutClose
This change increases the timeout for the connection state change
verifications in
`GattClientTest#connectAndDisconnectManyClientsWithoutClose` from 1
second to 5 seconds. The test has been observed to be flaky,
with a ~13% failure rate. This increase in the timeout allows for more
time for the status change to be confirmed.

Local runs show that while the timeouts are longer, the average test
execution time remains consistent with the previous timeouts,
but the test now has a 100% pass rate. This indicates that the original
timeout was too aggressive and was the cause of the intermittent
failures.

Bug: 404652736
Fix: 441838424
Test: atest BumbleBluetoothTests:android.bluetooth.GattClientTest#connectAndDisconnectManyClientsWithoutClose
Flag: EXEMPT TEST_ONLY
Change-Id: Ibca6b46ab8c6f7afa345e7df97aa03f9b37a097a